backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
langchain-deepseek
langchain-ollama
langchain-community
langchain-experimental
requests
beautifulsoup4
html2text
tavily-python
llama-index-core
llama-index-embeddings-openai
llama-index-retrievers-bm25
rank-bm25
pypdf
